Allow combination of flex and continuous stopping under certain conditions
