+++ application/src/ext/java/org/opentripplanner/ext/flex/trip/UnscheduledTrip.java
@@ -1,6 +1,5 @@
 package org.opentripplanner.ext.flex.trip;
 
-import static org.opentripplanner.model.PickDrop.NONE;
 import static org.opentripplanner.model.StopTime.MISSING_VALUE;
 
 import java.util.Arrays;
@@ -8,7 +7,6 @@ import java.util.List;
 import java.util.Objects;
 import java.util.Optional;
 import java.util.Set;
-import java.util.function.Predicate;
 import java.util.stream.Collectors;
 import org.opentripplanner.ext.flex.flexpathcalculator.FlexPathCalculator;
 import org.opentripplanner.ext.flex.flexpathcalculator.TimePenaltyCalculator;
@@ -81,11 +79,9 @@ public class UnscheduledTrip extends FlexTrip<UnscheduledTrip, UnscheduledTripBu
    *  - One or more stop times with a flexible time window but no fixed stop in between them
    */
   public static boolean isUnscheduledTrip(List<StopTime> stopTimes) {
-    Predicate<StopTime> hasContinuousStops = stopTime ->
-      stopTime.getFlexContinuousDropOff() != NONE || stopTime.getFlexContinuousPickup() != NONE;
     if (stopTimes.isEmpty()) {
       return false;
-    } else if (stopTimes.stream().anyMatch(hasContinuousStops)) {
+    } else if (stopTimes.stream().anyMatch(StopTime::combinesContinuousStoppingWithFlex)) {
       return false;
     } else if (N_STOPS.contains(stopTimes.size())) {
       return stopTimes.stream().anyMatch(StopTime::hasFlexWindow);
