+++ application/src/main/java/org/opentripplanner/model/StopTime.java
@@ -1,6 +1,8 @@
 /* This file is based on code copied from project OneBusAway, see the LICENSE file for further information. */
 package org.opentripplanner.model;
 
+import static org.opentripplanner.model.PickDrop.NONE;
+
 import java.util.List;
 import org.opentripplanner.framework.i18n.I18NString;
 import org.opentripplanner.transit.model.site.StopLocation;
@@ -304,5 +306,17 @@ public final class StopTime implements Comparable<StopTime> {
    */
   public boolean hasFlexWindow() {
     return flexWindowStart != MISSING_VALUE || flexWindowEnd != MISSING_VALUE;
+  }
+
+  /**
+   * Checks if this stop time combines flex windows with continuous stopping, which is against the
+   * GTFS spec.
+   */
+  public boolean combinesContinuousStoppingWithFlex() {
+    return hasContinuousStopping() && hasFlexWindow();
+  }
+
+  public boolean hasContinuousStopping() {
+    return this.flexContinuousPickup != NONE || flexContinuousDropOff != NONE;
   }
 }
